Update the way usernames and tags and other parts of the command are handled. This should fix it to work with both spaced names and unspaced names as well as optional queueries like size, mode, etc. However, it is nearly 1 am and I may have missed some tests and will probably need to look/test it further Added !match command and formatting Changed the way embed are handled by using global variables and setting the global variable's embed. Need to look into a way to do this with buttons/if it is possible Need to look into formatting of some embeds to see if I can line them all up etc.
!IMPORTANT!
In !match command change to lifetime to be able to access much more matches and for !match history as well so it isnt limited at 10
!IMPORTANT!

Big chunky commit but getting there. Have about all the commands I want just need to clean up the formatting of all the command and maybe create functions to clean up code for things like username and tag processing etc.